fix(mlir): MSVC-signed C-enum bindings — u32 casts the comment promised but never applied (T0194)
Both Windows legs of the rolling dev release failed at `cargo build`
with six E0308s in verum_mlir: bindgen renders plain C enums as `i32`
on MSVC (they are signed there) and `u32` elsewhere, so the
`#[repr(u32)]` WalkOrder/WalkResult discriminants and the
`UnknownDiagnosticSeverity(u32)` constructor type-mismatched only on
Windows. The file's own comment already prescribed `as u32` per
constant — the casts were never written. Values are small and
non-negative, so every cast is exact on every platform; unix builds
stay warning-clean under -D warnings.

With this, the LLVM-prebuilt download leg (verified working in run
30892367677: "Downloaded via gh release download") should carry both
Windows legs to completion — the 4/6 dev assets today are the two
macOS + two Linux triples; this was the last blocker for 6/6.

+++ crates/llvm/verum_mlir/src/diagnostic/severity.rs
@@ -31,7 +31,11 @@ impl TryFrom<MlirDiagnosticSeverity> for DiagnosticSeverity {
             MlirDiagnosticSeverity_MlirDiagnosticNote => Self::Note,
             MlirDiagnosticSeverity_MlirDiagnosticRemark => Self::Remark,
             MlirDiagnosticSeverity_MlirDiagnosticWarning => Self::Warning,
-            _ => return Err(Error::UnknownDiagnosticSeverity(severity)),
+            // `as u32`: the bindgen alias is `i32` on MSVC (plain C
+            // enums are signed there) and `u32` elsewhere; severity
+            // codes are small and non-negative, so the cast is exact
+            // on every platform.
+            _ => return Err(Error::UnknownDiagnosticSeverity(severity as u32)),
         })
     }
 }
